eval/ruler/requirements.txt
torch==2.4.1
transformers==4.44.2
flask
flask_restful
sshtunnel_requests
tritonclient[all]
wonderwords
tenacity
huggingface-hub==0.23.2
html2text
google-generativeai
nltk
PyYAML
nemo_toolkit
hydra-core
pytorch_lightning
wonderwords
librosa
sentencepiece
youtokentome
editdistance
lhotse
pyannote.audio
webdataset
datasets
jiwer
IPython
bs4
inflect
accelerate